+++ src/domain/repositories/PaginatedSurveyRepository.ts
@@ -1,5 +1,5 @@
 import { FutureData } from "../../data/api-futures";
-import { ProgramCountMap } from "../entities/Program";
+import { SurveyChildCountType } from "../../data/utils/surveyChildCountHelper";
 import { Id } from "../entities/Ref";
 import { Survey, SURVEY_FORM_TYPES } from "../entities/Survey";
 import { PaginatedReponse } from "../entities/TablePagination";
@@ -17,12 +17,10 @@ export interface PaginatedSurveyRepository {
         keyword: string,
         orgUnitId: Id
     ): FutureData<PaginatedReponse<Survey[]>>;
-    getSurveyChildCount(
+    getPaginatedSurveyChildCount(
         parentProgram: Id,
         orgUnitId: Id,
         parentSurveyId: Id,
         secondaryparentId: Id | undefined
-    ):
-        | { type: "value"; value: FutureData<number> }
-        | { type: "map"; value: FutureData<ProgramCountMap> };
+    ): SurveyChildCountType;
 }
