+++ src/domain/repositories/SurveyRepository.ts
@@ -1,5 +1,6 @@
 import { FutureData } from "../../data/api-futures";
-import { ImportStrategy, ProgramCountMap } from "../entities/Program";
+import { SurveyChildCountType } from "../../data/utils/surveyChildCountHelper";
+import { ImportStrategy } from "../entities/Program";
 import { Questionnaire } from "../entities/Questionnaire/Questionnaire";
 import { Id } from "../entities/Ref";
 import { Survey, SURVEY_FORM_TYPES } from "../entities/Survey";
@@ -33,12 +34,10 @@ export interface SurveyRepository {
 
     getSurveyNameFromId(id: Id, surveyFormType: SURVEY_FORM_TYPES): FutureData<string>;
 
-    getSurveyChildCount(
+    getNonPaginatedSurveyChildCount(
         parentProgram: Id,
         orgUnitId: Id,
         parentSurveyId: Id,
         secondaryparentId: Id | undefined
-    ):
-        | { type: "value"; value: FutureData<number> }
-        | { type: "map"; value: FutureData<ProgramCountMap> };
+    ): SurveyChildCountType;
 }
