+++ src/webapp/components/survey-list/table/PaginatedSurveyListTable.tsx
@@ -19,13 +19,14 @@ import {
 import i18n from "@eyeseetea/feedback-component/locales";
 import { ActionMenuButton } from "../../action-menu-button/ActionMenuButton";
 import { palette } from "../../../pages/app/themes/dhis2.theme";
-import { Dispatch, MouseEventHandler, SetStateAction, useEffect, useState } from "react";
+import { Dispatch, SetStateAction, useEffect, useState } from "react";
 import { ArrowDownward, ArrowUpward } from "@material-ui/icons";
 import _ from "../../../../domain/entities/generic/Collection";
 import { useDeleteSurvey } from "../hook/useDeleteSurvey";
 import { ContentLoader } from "../../content-loader/ContentLoader";
 import { useSurveyListActions } from "../hook/useSurveyListActions";
 import { getChildrenName } from "../../../../domain/utils/getChildrenName";
+import { useMultipleChildCount } from "../hook/useMultipleChildCount";
 
 interface PaginatedSurveyListTableProps {
     surveys: Survey[] | undefined;
@@ -51,14 +52,6 @@ export const PaginatedSurveyListTable: React.FC<PaginatedSurveyListTableProps> =
     //states for column sort
     const [surveyNameSortDirection, setSurveyNameSortDirection] = useState<SortDirection>("asc");
     const [patientIdSortDirection, setPatientIdSortDirection] = useState<SortDirection>("asc");
-    const [sampleShipmentsSortDirection, setSampleShipmentsSortDirection] =
-        useState<SortDirection>("asc");
-    const [centralRefLabsResultsSortDirection, setCentralRefLabsResultsSortDirection] =
-        useState<SortDirection>("asc");
-    const [pathogenIsolatesLogsSortDirection, setPathogenIsolatesLogsSortDirection] =
-        useState<SortDirection>("asc");
-    const [supranationalRefsResultsSortDirection, setSupranationalRefsResultsSortDirection] =
-        useState<SortDirection>("asc");
 
     const { deleteSurvey, loading, deleteCompleteState } = useDeleteSurvey(
         surveyFormType,
@@ -76,6 +69,8 @@ export const PaginatedSurveyListTable: React.FC<PaginatedSurveyListTableProps> =
         sortByColumn,
     } = useSurveyListActions(surveyFormType);
 
+    const { getCurrentSortDirection, childOnClick } = useMultipleChildCount(sortByColumn);
+
     useEffect(() => {
         if (surveys) setSortedSurveys(surveys);
 
@@ -86,56 +81,6 @@ export const PaginatedSurveyListTable: React.FC<PaginatedSurveyListTableProps> =
             snackbar.error(deleteCompleteState.message);
         }
     }, [deleteCompleteState, snackbar, surveys, setSortedSurveys]);
-
-    const getCurrentSortDirection = (childOptionName: string): SortDirection => {
-        switch (childOptionName) {
-            case "Sample Shipment":
-                return sampleShipmentsSortDirection;
-            case "Central Ref Lab Results":
-                return centralRefLabsResultsSortDirection;
-            case "Pathogen Isolates Logs":
-                return pathogenIsolatesLogsSortDirection;
-            case "Supranational Ref Results":
-                return supranationalRefsResultsSortDirection;
-            default:
-                throw new Error(`Invalid child option name: ${childOptionName}`);
-        }
-    };
-
-    const childOnClick = (childOptionName: string): MouseEventHandler | undefined => {
-        switch (childOptionName) {
-            case "Sample Shipment":
-                return () => {
-                    sampleShipmentsSortDirection === "asc"
-                        ? setSampleShipmentsSortDirection("desc")
-                        : setSampleShipmentsSortDirection("asc");
-                    sortByColumn("childCount", sampleShipmentsSortDirection);
-                };
-            case "Central Ref Lab Results":
-                return () => {
-                    centralRefLabsResultsSortDirection === "asc"
-                        ? setCentralRefLabsResultsSortDirection("desc")
-                        : setCentralRefLabsResultsSortDirection("asc");
-                    sortByColumn("childCount", centralRefLabsResultsSortDirection);
-                };
-            case "Pathogen Isolates Logs":
-                return () => {
-                    pathogenIsolatesLogsSortDirection === "asc"
-                        ? setPathogenIsolatesLogsSortDirection("desc")
-                        : setPathogenIsolatesLogsSortDirection("asc");
-                    sortByColumn("childCount", pathogenIsolatesLogsSortDirection);
-                };
-            case "Supranational Ref Results":
-                return () => {
-                    supranationalRefsResultsSortDirection === "asc"
-                        ? setSupranationalRefsResultsSortDirection("desc")
-                        : setSupranationalRefsResultsSortDirection("asc");
-                    sortByColumn("childCount", supranationalRefsResultsSortDirection);
-                };
-            default:
-                return undefined;
-        }
-    };
 
     return (
         <ContentLoader loading={loading} error="" showErrorAsSnackbar={false}>
